Create MinFallingPathSum.java
This problem is similar to unique paths but we need to return the minimum or shortest path. Return minimum of right and down. If right or down is zero due to out of bounds, return which is not zero. The condition where both are zero won't occur since its the goal and its the base condition