Create CoinChange.java
This problem can be solved with Recusion. Pick and Not pick approach. But here, we won't increment the index after picking up the coin. As we can take as many coins as we want. The base conditions would be the same. i.e. Return 0 if target is 0. Return Integer.MAX_VALUE if target is less than zero and if index reaches length of the array. Return the minimum of pick and not pick. Also, add 1 to the pick int after each recursion call, since we have considered a coin. We can use 2D dp to solve this problem, since it requires target and index. Create the dp array of (size) * (target+1). Check for target value at the index in the dp array and return it.